[POLARIS] BE Fix : Trade id 자료형 Long으로 변경

+++ back-end/polaris/src/main/java/com/ssafy/polaris/chat/repository/TradeRepository.java
@@ -3,7 +3,7 @@ package com.ssafy.polaris.chat.repository;
 import com.ssafy.polaris.trade.domain.Trade;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TradeRepository extends JpaRepository<Trade, Integer> {
+public interface TradeRepository extends JpaRepository<Trade, Long> {
 	// 나의 채팅방 목록
 	// List<Trade> findAllByUser(User user);
 }
